Added: .pet dismiss GM command, thanks SleepCoder http://arcemu.org/forums/index.php?showtopic=14851 Fixed: Failed spell casts would interrupt mana regeneration. Fixed: Spells (51729, 51265, 52051) Blessing spells from the objects in Zul'Drak Fixed: Compiler warning in GuildHandler.cpp (size_t going to uint32 without explicit typecast)

+++ src/arcemu-world/GuildHandler.cpp
@@ -1742,7 +1742,7 @@ void WorldSession::HandleGuildBankQueryText( WorldPacket & recv_data )
 	if( tab == NULL )
 		return;
 
-	uint32 len = tab->szTabInfo != NULL ? strlen( tab->szTabInfo ) : 1;
+	uint32 len = tab->szTabInfo != NULL ? (uint32)strlen( tab->szTabInfo ) : 1;
 
 	WorldPacket data( MSG_QUERY_GUILD_BANK_TEXT, 1 + len );
 	data << tabid;
